Fix definition of wasm's long long for conversions

+++ include/rlbox_wasmtime_sandbox.hpp
@@ -230,7 +230,7 @@ rlbox_wasmtime_sandbox_thread_data* get_rlbox_wasmtime_sandbox_thread_data();
 class rlbox_wasmtime_sandbox
 {
 public:
-  using T_LongLongType = int32_t;
+  using T_LongLongType = int64_t;
   using T_LongType = int32_t;
   using T_IntType = int32_t;
   using T_PointerType = uint32_t;
